+++ Iliad/Sources/Utilities/Utility.swift
@@ -39,4 +39,19 @@ class Utility {
         let emailTest = NSPredicate(format: "SELF MATCHES %@", emailRegEx)
         return emailTest.evaluate(with: testStr)
     }
+
+    static func setUserDefaults(userDefaults: UserDefaults?, values: [String: Any?]) {
+        for value in values {
+            userDefaults?.setValue(value.value, forKey: value.key)
+        }
+    }
+
+
+    static func showAlert(title: String?, message: String?, okHandler: ((UIAlertAction) -> Void)? = nil, cancelHandler: ((UIAlertAction) -> Void)? = nil) {
+        let alert = UIAlertController(title: title, message: message, preferredStyle: .alert)
+
+        alert.addAction(UIAlertAction(title: "Commons" ~> "OK", style: .default, handler: okHandler))
+        alert.addAction(UIAlertAction(title: "Commons" ~> "CANCEL", style: .default, handler: cancelHandler))
+
+    }
 }