+++ Changelog.md
@@ -6,6 +6,27 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.1] - Build 7 - 2018-09-033
+### Added
+- Consumption extension
+- Autologin alert
+
+### Fixed
+- Fix cancel to back
+- Fix Biometric login and auto login
+
+## [1.0.0] - Build 6 - 2018-08-31
+### Added
+- Policy view
+
+### Fixed
+
+### Changed
+- New consumption interface
+- New profile interface
+
+### Removed
+
 ## [1.0.0] - Build 2 - 2018-08-18
 ### Added
 - Error haptic feedback in Login and Recover Password.